Moved composition-table to library, bump bio
bumped the Bio lib minor version
