+++ packages/Bio/src/widgets/composition-analysis-widget.ts
@@ -12,6 +12,7 @@ import {HelmTypes} from '@datagrok-libraries/bio/src/helm/consts';
 
 import '../../css/composition-analysis.css';
 import {ISeqHelper} from '@datagrok-libraries/bio/src/utils/seq-helper';
+import {buildCompositionTable} from '@datagrok-libraries/bio/src/utils/composition-table';
 
 export function getCompositionAnalysisWidget(
   val: DG.SemanticValue, monomerLib: IMonomerLibBase, seqHelper: ISeqHelper
@@ -43,41 +44,3 @@ export function getCompositionAnalysisWidget(
   return new DG.Widget(host);
 }
 
-export function buildCompositionTable(
-  counts: { [m: string]: number }, biotype: HelmType, monomerLib: IMonomerLibBase
-): HTMLTableElement {
-  let sumValue: number = 0;
-  let maxValue: number | null = null;
-  for (const value of Object.values(counts)) {
-    sumValue = sumValue + value;
-    maxValue = maxValue === null ? value : Math.max(maxValue, value);
-  }
-  const maxRatio = maxValue! / sumValue;
-  const elMap: { [m: string]: HTMLElement } = Object.assign({}, ...Array.from(Object.entries(counts))
-    .sort((a, b) => b[1] - a[1])
-    .map(([cm, value]) => {
-      const ratio = value / sumValue;
-      const wem = monomerLib.getWebEditorMonomer(biotype, cm)!;
-      const color = wem.backgroundcolor!;
-      const barDiv = ui.div('', {classes: 'macromolecule-cell-comp-analysis-bar'});
-      barDiv.style.width = `${50 * ratio / maxRatio}px`;
-      barDiv.style.backgroundColor = color;
-      if (GAP_SYMBOL === cm) {
-        barDiv.style.borderWidth = '1px';
-        barDiv.style.borderStyle = 'solid';
-        barDiv.style.borderColor = DG.Color.toHtml(DG.Color.lightGray);
-      }
-      const displayMonomer: string = GAP_SYMBOL === cm ? '-' : cm;
-      const valueDiv = ui.div(`${(100 * ratio).toFixed(2)}%`);
-      const el = ui.div([barDiv, valueDiv], {classes: 'macromolecule-cell-comp-analysis-value'});
-      return ({[displayMonomer]: el});
-    }));
-
-  const table = ui.tableFromMap(elMap);
-  Array.from(table.rows).forEach((row) => {
-    const barCol = (row.getElementsByClassName('macromolecule-cell-comp-analysis-bar')[0] as HTMLDivElement)
-      .style.backgroundColor;
-    row.cells[0].style.color = barCol;
-  });
-  return table;
-}
